Move native-only issue reporter editor files to electron-browser
The issue reporter editor pane and its keybindings are only registered
and loaded by the electron-browser contribution, but lived under browser/
and reached into the native environment service for tmpDir behind an
eslint-disable for the layering and import-pattern rules.

Move both files to electron-browser/ where the native environment service
import is legal, removing the suppression. The wizard overlay and editor
input stay in browser/ since they are layer-agnostic.
